requirements.txt
Wikidata
python-gedcom
argparse
six